Refactor GitHub API to authenticate on instantiation
This commit changes the GitHub class to require authentication credentials
via OAuth or a Personal Access Token upon instantiation. The previous
API required you to call an authentication method first before calling
any of the other methods (which could potentially cause issues).

As we will require authentication credentials at instantiation, there is
no longer any need to have any authentication methods on the class, with
the exception of the `requestOAuthAccessToken` method. This is a static
method which *could* be refactored out of the class (to be determined later).

Some additional polish in the tests has also been performed.