Add new authentication handling for processing entries
This commit adds new handling for when entries are processed with the
new `auth.required` config option set to true. Staticman will now check
for an `auth-token` field (instead of a `github-token`) and use this to
validate that the user is indeed authenticated already on either GitHub
or GitLab.

Backwards compatibility with the v2 API has been maintained but it should
be analyzed if this is important as the code is somewhat messy as a result.