Speed up tests by using `cloneDeep` for site config
This commit speeds up the tests by up to ~4x (from ~48s to 12s). The
slowness was due to the high cost of instantiating Convict from the site
config every time a test was run. This is mostly due to all the
validations/coercions that it must perform (such as RSA decryption)
every single time. It is much faster to just clone the site config object.

- `lodash` was used for the `cloneDeep` implementation.