Add `auth-type` option when requiring auth and separate out v2 auth logic
This commit adds a new `auth-type` option that can be passed through to
the `/entry` endpoint when auth is required. This option allows the user
to choose which auth provider they wish to authenticate against i.e.
GitHub or GitLab, independently from the `service` that has been selected
in the endpoint URL.

Additionally, `_checkAuth` has been refactored to move out v2 into a
separate method to clean up the code. Previously there were too many
conditionals on the version number which made it harder to see what was
going on. The tests have also been cleaned up and consolidated.