Add `provider` query parameter for `/auth` endpoint
This commit adds the `provider` query parameter to allow users to select
the specific authentication provider they wish to use when authenticating
through Staticman.

This is useful in situations where the site repository is hosted on a
different provider to the authentication provider. For example, if the
site is hosted on GitHub, but the commenter wishes to sign in through
GitLab. Consequently, this is a prerequisite to adding additional
authentication providers in the future.