Add Heroku instructions, remove public instance refs, fix app auth.
Instructions for creating a private instance hosted on the free tier of
Heroku have been added and all references to the public instance have
been removed. Heroku hosting is intended to provide users with a more
reliable alternative to the public instance, which we will be dropping
support for.

Many changes had to be made to enable authentication as a GitHub
application. This required a lot of methods to become async.