Fix GitHub auth, upgrade GitHub API module, allow test suite to execute. (#319)
This is a large change meant to restore the repo to a functioning state. The GitHub library was reworked to use the lastest version of the GitHub API module and changes were made to support authentication as a GitHub application.

Instructions for creating a private instance hosted on the free tier of
Heroku have been added and all references to the public instance have
been removed. Heroku hosting is intended to provide users with a more
reliable alternative to the public instance, which we will be dropping
support for.